Avoid passing preinstalled jars to proguard and allow disabling the cache.
Proguard breaks if it is passed jars that are already part of the android
library. I copied the preinstalled list from jberkel/android-plugin.

Also, if proguardCache in Android is set to Seq() it will now disable the
broken caching stuff. I'll debug it more later, but it seems to be
subproject related. It breaks when it tries to open a classes/ dir as if
it were a .jar file.